requirements.txt
requests==2.31.0
python-dotenv==1.0.0
googletrans==3.1.0a0
